// Web/RMA-Member-Portal/memberportal-app/src/app/shared/enums/document-set.enum.ts
export enum DocumentSetEnum {

    FuneralNaturalIndividual = 1,
    Product = 2,
    ProductOption = 3,
    ProductBenefit = 4,
    FuneralUnnaturalSuicideIndividual = 5,
    FuneralStillBorn = 6,
    FuneralNaturalGroup = 7,
    FuneralUnnaturalSuicideGroup = 8,
    Brokerage = 9,
    NaturalRepresentative = 10,
    JuristicRepresentative = 11,
    PolicyCaseIndividual = 12,
    PolicyCaseGroup = 13,
    PolicyMaintanance = 14,
    PolicyCancellation = 15,
    MemberRelations = 16,
    MoveBrokerPolicies = 17,
    PolicyDocumentsindividual = 18,
    PolicyDocumentsgroup = 19,
    MissedPremiumLetter = 20,
    PolicyReinstatement = 21,
    PolicyContinuation = 22,
    ClaimRecovery = 23,
    MaintainRefundOverpayment = 24,
    CreditNoteReallocation = 25,
    EmailAttachment = 26,
    Unclaimbenefit = 27,
    InvestigationDocuments = 28,
    InterBankTransfer = 29,
    TraceDocument = 30,
    UserRegistrationPassportDocuments = 31,
    UserVopdFailedDocuments = 32,
    GroupPolicyMember = 33,
    LeadsDocuments = 34,
    AcceptedQuoteDocuments = 35,
    MemberDocumentSet = 36,
    LapsePolicyDocumentSet = 37,
    ClaimAccidentNotificationID = 38,
    CoidPolicyCancellation = 39,
    ClaimDiseaseNotificationID = 40,
    ClaimAccidentNotificationPassport = 41,
    ClaimDiseaseNotificationPassport = 42,
    DigiCareFirstMedicalReport = 43,
    TermsArrangementDocuments = 44,
    PensCareFirstMedicalReport = 45,
    PolicyStatusChange = 46,
    PensionCaseDocuments = 51,
    DigiCareProgressMedicalReport = 52,
    DigiCareFinalMedicalReport = 53,
    MovePolicyScheme = 54,
    Declaration = 55,
    CoidPolicyRefund = 56,
    RateAdjustment = 57,
    ChildExtensionDocuments = 58,
    UpgradeDowngradePolicy = 59,
    CertificateOfLifeVendorDocuments = 60,
    ConsolidatedFuneralPolicyDocuments = 61,
    PensionIncrease = 62,
    PensionChildExtension = 63,
    UnsuccessfulCommutation = 64,
    SuccessfulCommutationletter = 65,
    ChildTurnsEighteen = 66,
    LedgerToExpire = 67,
    PolicyDocuments = 68,
    IdentificationDocuments = 69,
    NIHL = 70,
    HeatClaimsDocuments = 72,
    WRULDdocuments = 73,
    MalariaDocuments = 74,
    TuberculosisDocuments = 76,
    CovidDocuments = 77,
    PtsdDocuments = 78,
    AsthmaDocuments = 79,
    NIHLRequireRetest = 83,
    CommonPersonalDocuments = 86,
    ClaimDiseaseDocuments = 87,
    ClaimMedicalDocuments = 88,
    ClaimsAdditionalDocuments = 89,
    OptionalPersonalDocuments = 90,
    TermsSupportingDocuments = 91,
    LettersOfGoodStanding = 92,
    PolicyAmendment = 93,
    VopdManualVerification = 94,
    Remittance = 95,
    RequestingAdditionalDocuments = 96,
    MedicalProstheticDocuments = 97,
    MedicalProstheticOrthoticDocsOver2YrsNonPension = 98,
    MedicalOrthoticDocsOver2YrsNonPension = 99,
    MedicalProstheticDocsOver2YrsPension = 100,
    LegalObjectionDocumentPack = 101,
    LegalRecoveryDocumentPack = 102,
    PulmonaryTuberculosisDocuments = 103,
    ClaimRequirementsDocuments = 104,
    TtdDocuments = 105,
    AssaultDocuments = 106,
    MVA = 107,
    RIOT = 108,
    BankStatement = 109,
    InvestigationReferralDocuments = 110,
    InterdebtorTransfer = 111,
    MedicalChronicDocuments = 112,
    AdohcDebitOrderDocuments = 113,
    MedicalProstheticPhotos = 114,
    MedicalProstheticCoidForm = 115,
    EmployeeEarningsDocuments = 116,
    RefundTermsOverPayment = 117,
    RefundPolicyCancellation = 118,
    RefundCreditAdjustment = 119,
    ReferralSupportingDocuments = 120,
    GroupRisk = 121,
    HealthcareProviderRegistrationDocuments = 122,
    Athsma = 123,
    PulmonaryTB = 124,
    Dermatitis = 125,
    Fatal = 126,
    CJP = 127,
    LetterheadConfirmingDetails,
}