// Web/RMA-Member-Portal/memberportal-app/src/app/shared/enums/document-system-name-enum.ts
export enum DocumentSystemNameEnum {
  CommonManager = 1,
  ProductManager = 2,
  LeadManager = 3,
  QuoteManager = 4,
  MemberManager = 5,
  PolicyManager = 6,
  ClaimManager = 7,
  BillingManager = 8,
  PensCareManager = 9,
  ChildExtensionManager = 10,
  WizardManager = 11,
  LegalCareManager = 12,
  DebtCareManager = 13,
  MediCareManager = 14,
  RolePlayerDocuments = 15
}