// Web/RMA-Member-Portal/memberportal-app/src/app/shared/enums/health-care-provider-group-enum.ts
export enum HealthCareProviderGroupEnum {
    Unknown = 1000,
    MediClinic = 1001,
    CLINIXHOSPITALS = 1002,
    LIFEHEALTHCARE = 1003,
    MineHospitals = 1004,
    NHN = 1006,
    NAMIBIANHOSPITALS = 1007,
    NETCARE = 1008,
    PUBLICSECTOR = 1009,
    CARECUREGROUP = 1010,
    COMPSOL = 1011,
    COIDLINK = 1012,
    MEDICROSS = 1013,
    SPESNET = 1014,
    WORKABILITY = 1015,
    MEDSOL = 1016,
    VanHeerdenGroup = 1017,
    Partner4life = 1018,
    BUSAMED = 1019,
    MBSCC = 1020,
    DDC = 1021,
    ICW = 1022,
    SummitMe = 1023,
    Denticall = 1024,
    VProfessional = 1025,
    GoodPrognosis = 1026,
    AdminestratorsMCM = 1027,
    CLBC = 1028,
    ASJBilling = 1029,
    LenMed = 1030
}